fix fetch cities hook
